reformatted according to Maven coding standard

+++ Common/Util-All/src/main/java/ch/jfactory/component/WaitOverlay.java
@@ -50,7 +50,6 @@ import javax.swing.Timer;
  */
 public class WaitOverlay extends JComponent
 {
-
     private final int width;
 
     private final int height;
@@ -69,35 +68,35 @@ public class WaitOverlay extends JComponent
 
     private final Color overlayColor;
 
-    public WaitOverlay(final JDialog dialog, final Color overlayColor)
+    public WaitOverlay( final JDialog dialog, final Color overlayColor )
     {
         this.width = dialog.getWidth();
         this.height = dialog.getHeight();
-        dialog.setGlassPane(this);
+        dialog.setGlassPane( this );
         this.overlayColor = overlayColor;
     }
 
     public void start()
     {
         running = true;
-        timer = new Timer(30, new ActionListener()
+        timer = new Timer( 30, new ActionListener()
         {
-            public void actionPerformed(final ActionEvent e)
+            public void actionPerformed( final ActionEvent e )
             {
                 advance();
                 ticker.tick();
             }
-        });
-        timer.setRepeats(true);
+        } );
+        timer.setRepeats( true );
         timer.start();
-        setVisible(true);
+        setVisible( true );
     }
 
     public void stop()
     {
-        if (!running)
+        if ( !running )
         {
-            throw new IllegalStateException("Not running");
+            throw new IllegalStateException( "Not running" );
         }
         running = false;
         timer.stop();
@@ -109,7 +108,7 @@ public class WaitOverlay extends JComponent
      *
      * @param alpha the opacity
      */
-    public void setAlpha(final float alpha)
+    public void setAlpha( final float alpha )
     {
         this.alpha = alpha;
     }
@@ -121,10 +120,10 @@ public class WaitOverlay extends JComponent
 
     private void advance()
     {
-        if (ticker == null || ticker.isDone())
+        if ( ticker == null || ticker.isDone() )
         {
             state += 1;
-            if (state < tickers.length)
+            if ( state < tickers.length )
             {
                 ticker = tickers[state];
                 ticker.start();
@@ -133,27 +132,26 @@ public class WaitOverlay extends JComponent
     }
 
     @Override
-    protected void paintComponent(final Graphics g)
+    protected void paintComponent( final Graphics g )
     {
-        super.paintComponent(g);
+        super.paintComponent( g );
         final Graphics2D g2 = (Graphics2D) g;
-        g2.setColor(overlayColor);
-        g2.setComposite(AlphaComposite.getInstance(AlphaComposite.SRC_OVER, 1f - alpha));
-        g2.fillRect(0, 0, width, height);
-        if (ticker != null)
+        g2.setColor( overlayColor );
+        g2.setComposite( AlphaComposite.getInstance( AlphaComposite.SRC_OVER, 1f - alpha ) );
+        g2.fillRect( 0, 0, width, height );
+        if ( ticker != null )
         {
-            ticker.paint(g2);
+            ticker.paint( g2 );
         }
     }
 
-    public void setTickers(final Ticker... tickers)
+    public void setTickers( final Ticker... tickers )
     {
         this.tickers = tickers;
     }
 
     public abstract class Ticker
     {
-
         protected long start;
 
         protected WaitOverlay waitOverlay;
@@ -167,9 +165,9 @@ public class WaitOverlay extends JComponent
             resetTime();
         }
 
-        protected float getPercent(final float time)
+        protected float getPercent( final float time )
         {
-            return Math.min(1f, (float) (System.currentTimeMillis() - start) / (float) time);
+            return Math.min( 1f, (float) ( System.currentTimeMillis() - start ) / time );
         }
 
         protected void resetTime()
@@ -186,19 +184,18 @@ public class WaitOverlay extends JComponent
         {
         }
 
-        public void paint(final Graphics2D g2)
+        public void paint( final Graphics2D g2 )
         {
         }
     }
 
     public final class FadeInTicker extends Ticker
     {
-
         private final float dimTime;
 
         private final float endAlpha;
 
-        public FadeInTicker(final float endAlpha, final float dimTime)
+        public FadeInTicker( final float endAlpha, final float dimTime )
         {
             this.dimTime = dimTime;
             this.endAlpha = endAlpha;
@@ -206,20 +203,19 @@ public class WaitOverlay extends JComponent
 
         public void tick()
         {
-            final float alpha = 1f + (endAlpha - 1f) * getPercent(dimTime);
-            setAlpha(alpha);
+            final float alpha = 1f + ( endAlpha - 1f ) * getPercent( dimTime );
+            setAlpha( alpha );
             WaitOverlay.this.getParent().repaint();
         }
 
         public boolean isDone()
         {
-            return getPercent(dimTime) >= 1f;
+            return getPercent( dimTime ) >= 1f;
         }
     }
 
     public final class WaitForThreadTicker extends Ticker
     {
-
         private static final String IMAGE = "/ch/jfactory/component/wait-circle.png";
 
         private final BufferedImage waitIcon;
@@ -232,15 +228,15 @@ public class WaitOverlay extends JComponent
 
         private boolean joined;
 
-        public WaitForThreadTicker(final JComponent component, final Thread thread) throws IOException
+        public WaitForThreadTicker( final JComponent component, final Thread thread ) throws IOException
         {
             this.thread = thread;
             this.component = component;
-            final BufferedImage image = ImageIO.read(WaitForThreadTicker.class.getResourceAsStream(IMAGE));
+            final BufferedImage image = ImageIO.read( WaitForThreadTicker.class.getResourceAsStream( IMAGE ) );
             final AffineTransform at = new AffineTransform();
-            at.scale(2.0, 2.0);
-            final BufferedImageOp op = new AffineTransformOp(at, AffineTransformOp.TYPE_NEAREST_NEIGHBOR);
-            waitIcon = op.filter(image, null);
+            at.scale( 2.0, 2.0 );
+            final BufferedImageOp op = new AffineTransformOp( at, AffineTransformOp.TYPE_NEAREST_NEIGHBOR );
+            waitIcon = op.filter( image, null );
         }
 
         public void tick()
@@ -259,7 +255,7 @@ public class WaitOverlay extends JComponent
                     {
                         thread.join();
                     }
-                    catch (InterruptedException e)
+                    catch ( InterruptedException e )
                     {
                         e.printStackTrace();
                     }
@@ -274,55 +270,53 @@ public class WaitOverlay extends JComponent
         }
 
         @Override
-        public void paint(final Graphics2D g2)
+        public void paint( final Graphics2D g2 )
         {
             final int imageWidth = waitIcon.getWidth();
             final int imageHeight = waitIcon.getHeight();
             final int componentWidth = component.getWidth();
             final int componentHeight = component.getHeight();
-            final float millisPassed = ((float) (System.currentTimeMillis() - start)) % millisPerRotation;
+            final float millisPassed = ( (float) ( System.currentTimeMillis() - start ) ) % millisPerRotation;
             final float millisPerSegment = millisPerRotation / 12;
-            final int segmentsPassed = (int) (millisPassed / millisPerSegment);
-            g2.rotate(segmentsPassed * Math.PI / 6.0, (double) componentWidth / 2.0, (double) componentHeight / 2.0);
-            g2.drawImage(waitIcon, (componentWidth - imageWidth) / 2, (componentHeight - imageHeight) / 2, null);
+            final int segmentsPassed = (int) ( millisPassed / millisPerSegment );
+            g2.rotate( segmentsPassed * Math.PI / 6.0, (double) componentWidth / 2.0, (double) componentHeight / 2.0 );
+            g2.drawImage( waitIcon, ( componentWidth - imageWidth ) / 2, ( componentHeight - imageHeight ) / 2, null );
         }
     }
 
     public final class FadeOutTicker extends Ticker
     {
-
         private final float dimTime;
 
         private float startAlpha = -1;
 
-        public FadeOutTicker(final float dimTime)
+        public FadeOutTicker( final float dimTime )
         {
             this.dimTime = dimTime;
         }
 
         public void tick()
         {
-            if (startAlpha == -1)
+            if ( startAlpha == -1 )
             {
                 startAlpha = getAlpha();
             }
-            final float alpha = startAlpha + (1f - startAlpha) * getPercent(dimTime);
-            setAlpha(alpha);
+            final float alpha = startAlpha + ( 1f - startAlpha ) * getPercent( dimTime );
+            setAlpha( alpha );
             WaitOverlay.this.getParent().repaint();
         }
 
         public boolean isDone()
         {
-            return getPercent(dimTime) >= 1f;
+            return getPercent( dimTime ) >= 1f;
         }
     }
 
     public final class DisposeDialogTicker extends Ticker
     {
-
         private final JDialog dialog;
 
-        public DisposeDialogTicker(final JDialog dialog)
+        public DisposeDialogTicker( final JDialog dialog )
         {
             this.dialog = dialog;
         }
